+++ src/com/reason/lang/core/psi/impl/RPsiDuneVar.java
@@ -1,11 +1,12 @@
 package com.reason.lang.core.psi.impl;
 
-import com.intellij.lang.ASTNode;
-import com.reason.lang.dune.DuneTypes;
-import org.jetbrains.annotations.NotNull;
+import com.intellij.psi.tree.*;
+import com.reason.lang.core.type.*;
+import com.reason.lang.dune.*;
+import org.jetbrains.annotations.*;
 
-public class RPsiDuneVar extends RPsiToken<DuneTypes> {
-  public RPsiDuneVar(@NotNull DuneTypes types, @NotNull ASTNode node) {
-    super(types, node);
-  }
+public class RPsiDuneVar extends ORCompositePsiElement<DuneTypes> {
+    public RPsiDuneVar(@NotNull DuneTypes types, @NotNull IElementType elementType) {
+        super(types, elementType);
+    }
 }