+++ tests/com/reason/lang/dune/DuneVarParsingTest.java
@@ -1,19 +1,14 @@
 package com.reason.lang.dune;
 
-import com.intellij.psi.PsiElement;
-import com.reason.lang.BaseParsingTestCase;
+import com.intellij.psi.*;
 import org.junit.*;
 
-public class DuneVarParsingTest extends BaseParsingTestCase {
-  public DuneVarParsingTest() {
-    super("", "", new DuneParserDefinition());
-  }
+public class DuneVarParsingTest extends DuneParsingTestCase {
+    @Test
+    public void test_basic() {
+        PsiElement e = firstElement(parseRawCode("%{x}"));
 
-  @Test
-  public void test_basic() {
-    PsiElement e = firstElement(parseRawCode("%{x}"));
-
-    assertEquals(DuneTypes.INSTANCE.C_VAR, e.getNode().getElementType());
-    assertEquals("%{x}", e.getText());
-  }
+        assertEquals(DuneTypes.INSTANCE.C_VAR, e.getNode().getElementType());
+        assertEquals("%{x}", e.getText());
+    }
 }